pyproject.toml
[tool.poetry]
name = "rlfinitegames"
version = "0.1.0"
description = ""
authors = ["aferdina <aferdina@mail.uni-mannheim.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
gym = "^0.26.2"
matplotlib = "^3.7.1"
dash = "^2.9.1"
dash-bootstrap-components = "^1.4.1"
dash-mantine-components = "^0.12.0"
dash-iconify = "^0.1.2"
dash-extensions = "^0.1.13"
python-i18n = {extras = ["yaml"], version = "^0.3.9"}
loguru = "^0.6.0"
attr = "^0.3.2"
trio = "^0.22.0"
colorlog = "^6.7.0"


[tool.poetry.group.dev.dependencies]
pylint = "^2.17.0"
autopep8 = "^2.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
